fix: Change svalue_t::subtype from unsigned short to unsigned int
Fixes #1127

## Problem
When iterating over extremely large arrays or mappings (>65535 elements)
using foreach, an exception occurs due to type mismatch. The foreach
bytecode implementation stores the iteration counter in svalue_t::subtype,
which was defined as unsigned short (max value 65535). When array/mapping
size (stored as int) is assigned to subtype, values larger than 65535
overflow, causing iteration failures.

## Solution
Changed svalue_t::subtype from unsigned short (16-bit) to unsigned int
(32-bit). This allows foreach loops to handle collections with up to
4.2 billion elements without overflow.

## Memory Impact
None. Due to struct alignment padding, svalue_t remains 16 bytes on 64-bit
systems both before and after this change.

## Safety Analysis
- All bitwise operations on subtype use small flag constants (0x1, 0x2, 0x4)
  which work correctly with unsigned int
- No binary serialization directly accesses subtype field
- All 212 assignments to subtype throughout codebase are compatible

+++ src/vm/internal/base/svalue.h
@@ -38,7 +38,7 @@ union u {
  */
 struct svalue_t {
   unsigned short type;
-  unsigned short subtype;
+  unsigned int subtype;  // Changed from unsigned short to support large arrays/mappings in foreach (issue #1127)
   union u u;
 };
